Test spliting
